Refactor HNSW index into modular structure
- Split `src/index/vector/hnsw.rs` into `src/index/vector/hnsw/` directory
- Extracted `config`, `persistence`, `stats` modules
- Broke circular dependency between `HnswIndexBuilder` and `HnswIndex`
- Preserved full test suite in `tests.rs`
